+++ packages/core/src/components/fields/ObjectField.tsx
@@ -202,15 +202,18 @@ class ObjectField<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends Fo
     const newFormData = { ...formData } as T;
 
     let type: RJSFSchema['type'] = undefined;
+    let constValue: RJSFSchema['const'] = undefined;
     let defaultValue: RJSFSchema['default'] = undefined;
     if (isObject(schema.additionalProperties)) {
       type = schema.additionalProperties.type;
+      constValue = schema.additionalProperties.const;
       defaultValue = schema.additionalProperties.default;
       let apSchema = schema.additionalProperties;
       if (REF_KEY in apSchema) {
         const { schemaUtils } = registry;
         apSchema = schemaUtils.retrieveSchema({ $ref: apSchema[REF_KEY] } as S, formData);
         type = apSchema.type;
+        constValue = apSchema.const;
         defaultValue = apSchema.default;
       }
       if (!type && (ANY_OF_KEY in apSchema || ONE_OF_KEY in apSchema)) {
@@ -219,8 +222,9 @@ class ObjectField<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends Fo
     }
 
     const newKey = this.getAvailableKey('newKey', newFormData);
+    const newValue = constValue ?? defaultValue ?? this.getDefaultValue(type);
     // Cast this to make the `set` work properly
-    set(newFormData as GenericObjectType, newKey, defaultValue ?? this.getDefaultValue(type));
+    set(newFormData as GenericObjectType, newKey, newValue);
 
     onChange(newFormData);
   };